Added functionality to the menu option that changes the density of the Biham-Middleton-Levine traffic model

+++ source/cellularautomata.h
@@ -45,7 +45,7 @@ class CellularAutomata
                                int width,
                                unsigned short color);
         //BML traffic model
-        double BML_density;
+        int BML_density; // 0-100
 
     public:
         int setType(int t);
@@ -82,6 +82,6 @@ class CellularAutomata
         int setInitialState(int state);
 
         int getBMLdensity();
-        int setBMLdensity(double BMLdensity);
+        int setBMLdensity(int BMLdensity);
 };
 #endif
